refactor: extract handleVisualKey into visual.ts
